[Pdm2-11] Move Characters inside Players directory
