+++ src/app/(Restricted)/(Player)/Characters/components/CharacterItem.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { View } from "react-native";
-import Button from "../../../../common/components/Button";
+import Button from "../../../../../common/components/Button";
 import { formatDescription } from "../services/subtitleService";
 
 type CharacterData = {
