+++ src/app/(Restricted)/(Player)/Characters/components/CharacterList.tsx
@@ -2,7 +2,7 @@ import { FlatList } from "react-native";
 import React from "react";
 import AddCharacter from "./AddCharacter";
 import { CharacterItem } from "./CharacterItem";
-import { characters } from "../../../../mocks/MOCK_CHARACTER";
+import { characters } from "../../../../../mocks/MOCK_CHARACTER";
 
 export default function CharacterList() {
   return (
