+++ src/app/(Restricted)/(Player)/Characters/components/CharactersContent.tsx
