+++ src/app/(Restricted)/(Player)/Characters/screens/CharactersScreen.tsx
@@ -1,10 +1,10 @@
 import { View } from "react-native";
 import React from "react";
-import Background from "../../../../common/components/Background";
-import Watermark from "../../../../common/components/Watermark";
-import Footer from "../../../../common/components/Footer";
+import Background from "../../../../../common/components/Background";
+import Watermark from "../../../../../common/components/Watermark";
+import Footer from "../../../../../common/components/Footer";
 import CampaignsContent from "../components/CharactersContent";
-import Text from "../../../../common/components/Text";
+import Text from "../../../../../common/components/Text";
 
 export default function CharactersScreen() {
   return (
