+++ src/app/(Restricted)/(Player)/Characters/services/subtitleService.ts
