+++ src/app/_navigation/Player/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import CharactersScreen from "../../(Restricted)/Characters/screens/CharactersScreen";
+import CharactersScreen from "../../(Restricted)/(Player)/Characters/screens/CharactersScreen";
 
 export default function PlayerIndex() {
   return <CharactersScreen />;
